requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
assemblyai==0.32.0
python-dotenv==1.0.0
google-generativeai==0.7.2
pydantic==1.10.12
flask-sock==0.7.0
websocket-client==1.7.0
pydub==0.25.1
gunicorn==21.2.0
